requirements.txt
pathway
sentence-transformers
pandas
numpy
tqdm
openai
groq
python-dotenv
scikit-learn
torch
